refactor: reduce code duplication across codebase
Extract common helpers to reduce duplication from 7.39% to 6.1%:

- Extract toStream helper for AI providers (google, openrouter, cloudflare)
- Extract getPlanIdFromUrl helper for API route handlers
- Extract buildValidationErrorMessage for worker job handlers
- Extract common timestamp fields helper for database schemas
- Extract worker utility functions (sleep, normalizeError)

Impact:
- Created 5 new shared utility files
- Updated 16 existing files to use shared helpers
- Reduced TypeScript duplication from 7.39% to 6.84%
- Reduced overall duplication from 7.39% to 6.1%

Related to code quality improvements and DRY principles.
