+++ src/app/api/v1/plans/[planId]/attempts/route.ts
@@ -1,19 +1,14 @@
 import { withAuth, withErrorBoundary } from '@/lib/api/auth';
 import { NotFoundError, ValidationError } from '@/lib/api/errors';
 import { json } from '@/lib/api/response';
+import { getPlanIdFromUrl } from '@/lib/api/route-helpers';
 import { getPlanAttemptsForUser } from '@/lib/db/queries/plans';
 import { getUserByClerkId } from '@/lib/db/queries/users';
 import { mapAttemptsToClient } from '@/lib/mappers/detailToClient';
 
-function getPlanId(req: Request) {
-  const url = new URL(req.url);
-  const segments = url.pathname.split('/').filter(Boolean);
-  return segments[segments.length - 2];
-}
-
 export const GET = withErrorBoundary(
   withAuth(async ({ req, userId }) => {
-    const planId = getPlanId(req);
+    const planId = getPlanIdFromUrl(req, 'second-to-last');
     if (!planId) {
       throw new ValidationError('Plan id is required in the request path.');
     }
