+++ src/lib/ai/providers/openrouter.ts
@@ -13,23 +13,8 @@ import type {
   ProviderGenerateResult,
 } from '@/lib/ai/provider';
 import { PlanSchema } from '@/lib/ai/schema';
+import { toStream } from '@/lib/ai/utils';
 import { openRouterEnv } from '@/lib/config/env';
-
-function toStream(obj: unknown): AsyncIterable<string> {
-  const data = JSON.stringify(obj);
-  return {
-    [Symbol.asyncIterator](): AsyncIterator<string> {
-      let done = false;
-      return {
-        next(): Promise<IteratorResult<string>> {
-          if (done) return Promise.resolve({ done: true, value: undefined });
-          done = true;
-          return Promise.resolve({ done: false, value: data });
-        },
-      };
-    },
-  } as AsyncIterable<string>;
-}
 
 export interface OpenRouterProviderConfig {
   apiKey?: string; // OPENROUTER_API_KEY
