+++ src/lib/db/schema/tables/jobs.ts
@@ -12,6 +12,7 @@ import {
 } from 'drizzle-orm/pg-core';
 
 import { jobStatus, jobType } from '../../enums';
+import { timestampFields } from '../helpers';
 import { learningPlans } from './plans';
 import { users } from './users';
 import { authenticatedRole, clerkSub, serviceRole } from './common';
@@ -43,12 +44,7 @@ export const jobQueue = pgTable(
       .defaultNow(),
     startedAt: timestamp('started_at', { withTimezone: true }),
     completedAt: timestamp('completed_at', { withTimezone: true }),
-    createdAt: timestamp('created_at', { withTimezone: true })
-      .notNull()
-      .defaultNow(),
-    updatedAt: timestamp('updated_at', { withTimezone: true })
-      .notNull()
-      .defaultNow(),
+    ...timestampFields,
   },
   (table) => [
     check('attempts_check', sql`${table.attempts} >= 0`),
