+++ src/workers/plan-generator.ts
@@ -4,6 +4,7 @@ import { logger } from '@/lib/logging/logger';
 import { getNextJob } from '@/lib/jobs/queue';
 import { JOB_TYPES, type Job, type JobType } from '@/lib/jobs/types';
 import type { ProcessPlanGenerationJobResult } from './handlers/plan-generation-handler';
+import { normalizeError, sleep } from './utils';
 
 const DEFAULT_POLL_INTERVAL_MS = 2000;
 const DEFAULT_CONCURRENCY = 1;
@@ -36,26 +37,6 @@ export interface PlanGenerationWorkerStats {
   jobsCompleted: number;
   jobsFailed: number;
   [key: string]: unknown;
-}
-
-function sleep(ms: number): Promise<void> {
-  return new Promise((resolve) => setTimeout(resolve, ms));
-}
-
-function normalizeError(error: unknown): { message: string; name?: string } {
-  if (error instanceof Error) {
-    return { message: error.message, name: error.name };
-  }
-
-  if (typeof error === 'string' && error.length) {
-    return { message: error };
-  }
-
-  try {
-    return { message: JSON.stringify(error) };
-  } catch {
-    return { message: 'Unknown error' };
-  }
 }
 
 export class PlanGenerationWorker {
